backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# Redis
redis>=5.0.0

# InfluxDB
influxdb-client>=1.39.0

# Modbus
pymodbus>=3.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
numpy>=1.26.0
pyyaml>=6.0

# Development
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
